HOTT-1945: Dereference national measurement units in duty expressions
