+++ app/views/api/v1/measures/_measure_short.json.rabl
@@ -1,16 +1,16 @@
 attributes :id
 
-node(:vat) { |measure| measure.vat? }
+node(:vat, &:vat?)
 
-node(:measure_type) { |measure|
+node(:measure_type) do |measure|
   {
     id: measure.measure_type_id,
-    description: measure.measure_type.description
+    description: measure.measure_type.description,
   }
-}
-node(:duty_expression) { |measure|
+end
+node(:duty_expression) do |measure|
   {
-    base: measure.duty_expression_with_national_measurement_units_for(locals[:declarable]),
-    formatted_base: measure.formatted_duty_expression_with_national_measurement_units_for(locals[:declarable])
+    base: measure.duty_expression,
+    formatted_base: measure.formatted_duty_expression,
   }
-}
+end
